Fixed MSVC compiler flags, reverted some code regarding CString, fixed some warnings and removed a macro giving error while compiling in debug build.

+++ src/common/CScriptObj.cpp
@@ -1264,7 +1264,7 @@ badcmd:
 
 		case SSC_StrRegexNew:
 			{
-				size_t iLenString = Exp_GetVal( pszKey );
+				int iLenString = Exp_GetVal( pszKey );
 				tchar * sToMatch = Str_GetTemp();
 				if ( iLenString > 0 )
 				{
